(refactor) dbSeed & dbConnect created, functions working
