+++ server/index.js
@@ -1,19 +1,14 @@
 // require package to allow use of .env file
 require('dotenv').config();
+
 // require express for application server
 const express = require('express');
+
 // require package that service static front end files ./client/dist
 const path = require('path');
-// import dbConnection file
-require('./dbConnection');
-// // import the database connection function
-// const { connAndSeed } = require('./dbConfig')
-// // import the database DML functions
-// const { dummyData } = require('./dbFunctions')
 
-// // invoke oracle connection function and creat schema
-// connAndSeed();
-// dummyData();
+// import dbConnection file
+const { dbConnect } = require('./dbConnection');
 
 // set the port number for the application server
 const port = 3000
@@ -29,3 +24,6 @@ app.use(express.static(frontEnd))
 
 // use express listen function to invoke server instance, set to PORT for requests
 app.listen(port, () => console.info(`http://localhost:${port}`))
+
+// invoke oracle connection function and creat schema
+dbConnect();
